# YuGames.Documentation/YuGames API.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LeadOn's Corp - YuGames API",
    "description": "This API goal is to monitor players performance across multiple games.",
    "contact": {
      "name": "Valentin Virot",
      "email": "virot.valentin@gmail.com"
    },
    "version": "2.0"
  },
  "paths": {
...
    "/FifaGame/last/{number}/player/{playerId}": {
      "get": {
        "tags": ["FifaGame"],
        "summary": "Get last games played by player.",
        "description": "Get last games played by player with their team members.",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of games played.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FifaGamePlayedDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
...
    "/Platform": {
      "get": {
        "tags": ["Platform"],
        "summary": "Get all gaming platforms in database.",
        "responses": {
          "200": {
            "description": "Platform in database.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      },
      "post": {
        "tags": ["Platform"],
        "summary": "Creates a platform in database.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created platform.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Platform"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in."
          },
          "403": {
            "description": "User isn't admin."
          },
          "500": {
            "description": "Unknown error."
          }
        }
      },
      "patch": {
        "tags": ["Platform"],
        "summary": "Updates a platform in database.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Platform"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated platform.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Platform"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in."
          },
          "403": {
            "description": "User isn't admin."
          },
          "500": {
            "description": "Unknown error."
          }
        }
      }
    },
    "/Platform/{platformId}": {
      "get": {
        "tags": ["Platform"],
        "summary": "Get platform by ID.",
        "parameters": [
          {
            "name": "platformId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Platform.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Platform"
                }
              }
            }
          },
          "404": {
            "description": "Platform not found."
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
    "/Player/me": {
      "get": {
        "tags": ["Player"],
        "summary": "Gets current user profile.",
        "description": "Gets current user profile as stored in database. If the user account doesn't exists, it creates it automatically.",
        "responses": {
          "200": {
            "description": "Current user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      },
      "patch": {
        "tags": ["Player"],
        "summary": "Update current user profile.",
        "description": "Updates current user profile in database.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
    "/Player/{id}": {
      "get": {
        "tags": ["Player"],
        "summary": "Get a player by its ID.",
        "description": "Get a player by its ID, and retrieve its information.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player is found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "404": {
            "description": "Player not found."
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
    "/Player": {
      "get": {
        "tags": ["Player"],
        "summary": "Get all players in database.",
        "responses": {
          "200": {
            "description": "Players in database.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      },
      "patch": {
        "tags": ["Player"],
        "summary": "Update user profile (Admin only).",
        "description": "Updates user profile in database, only for admins.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
    "/Player/{playerId}/stats": {
      "get": {
        "tags": ["Player"],
        "summary": "Get all stats of a user.",
        "description": "Get all statistics for each platform of a user.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "seasonId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of stats.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FifaPlayerStatsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unknown error happened."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FifaGamePlayedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Player"
          },
          "playedOn": {
            "type": "string",
            "format": "date-time"
          },
          "platform": {
            "type": "string",
            "nullable": true
          },
          "platformId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isPlayed": {
            "type": "boolean"
          },
          "team1": {
            "$ref": "#/components/schemas/FifaTeamDto"
          },
          "team2": {
            "$ref": "#/components/schemas/FifaTeamDto"
          },
          "season": {
            "$ref": "#/components/schemas/Season"
          },
          "highlights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Highlight"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FifaPlayerStatsDto": {
        "type": "object",
        "properties": {
          "statsPerPlatform": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformStatsDto"
            },
            "nullable": true
          },
          "mostPlayedTeams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopTeamStatDto"
            },
            "nullable": true
          },
          "mostLossesTeams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopTeamStatDto"
            },
            "nullable": true
          },
          "mostWinsTeams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopTeamStatDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FifaTeam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FifaTeamDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "fifaTeamId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Highlight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdById": {
            "type": "integer",
            "format": "int32"
          },
          "fifaGameId": {
            "type": "integer",
            "format": "int32"
          },
          "externalUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Platform": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlatformStatsDto": {
        "type": "object",
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "draws": {
            "type": "integer",
            "format": "int32"
          },
          "averageGoalGiven": {
            "type": "number",
            "format": "float"
          },
          "averageGoalTaken": {
            "type": "number",
            "format": "float"
          },
          "goalsGiven": {
            "type": "integer",
            "format": "int32"
          },
          "goalsTaken": {
            "type": "integer",
            "format": "int32"
          },
          "goalDifference": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "float"
          },
          "looseRate": {
            "type": "number",
            "format": "float"
          },
          "drawRate": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "keycloakId": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "profilePictureUrl": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Season": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TopTeamStatDto": {
        "type": "object",
        "properties": {
          "numberOfGames": {
            "type": "integer",
            "format": "int32"
          },
          "fifaTeam": {
            "$ref": "#/components/schemas/FifaTeam"
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePlayerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "keycloakId": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "profilePictureUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}